Separated declarations and definitions.
Also, I'm adopting this new code style for macros where I use an extra backslash after the last line iff (not a typo!) the macro's contents does not require a semi-colon at the end. This should allow adding more to the end of the macro's definition/value on a new line.

...And I decided not to leave extra empty lines after this. Why? Because the compiler doesn't mind it, and managing when to use a single line or two lines is too much work!